# update HIP_UMA #7399