@jammm 2024-05-20
I would definitely benchmark this on Windows and Linux separately. I think it may perform slower on Windows.