@Djip007 2024-05-20
Is "HIP_UMA" possible on windows? 
I do not have windows to bench it sorry. But if it is possible I like to see what we can get (or not get...)

https://hipsolver.readthedocs.io/en/rocm-6.1.1/conceptual/gpu-memory.html#coherence
AMD suggest that if it can be use it must be use.