@Djip007 2024-05-27
> I have made some minor changes to the code to simplify it a bit and be more consistent, if everything looks good let's merge this.

Just have time to read it but looks good for me.